First draft of "Why We Created Julia" post.
This is an alternative attempt at the "Why Julia?" post.
I'm not sure if it goes over well at all or horribly, but
it's just an approach that came to me that got some flow
going while I was sitting at a cafe. Figured I'd write it
up and see how it went and how others like it. It's a bit
brash, but I think it kind of works.