Update Gaussian blur results to be correct for 18 threads.
This patch replaces the old Gaussian blur results figure, which had an
incorrect number for the 18-thread run, with a new figure generated
from results collected today.  The results are about the same as
before for the other numbers of threads.  Thanks to @StefanKarpinski
for being suspicious of the previous results.