Change Plot recommendations
Plots.jl is in the top 20 packages [according to the Julia Pulse (ranking via stars)](http://pkg.julialang.org/pulse.html). Because of its user-friendliness and high amount of development activity (with help always available in the Gitter channel), I believe it should be right here, front and center, on the Julia webiste. On the otherhand, most people will not be plotting directly into Compose.jl, and so I believe showing that as a "plotting package" confuses most newcommers (as evidenced by the questions we get in the Gitter/IRC about it. Plus the link was broken).

So I added a part for Plots.jl and removed the Part about Compose.jl. I think Plots.jl should come by default on JuliaBox to match this as well.