More Details in PDE Projects for GSoC
Given discussions in emails, I think it's wise to extend the parts on PDEs a little bit. I extended the "Discretizations of partial differential equations" project to give more details on what will be done, and directions the students can go. It's still a very open project, but this gives some very concrete directions.

Secondly, I added a project for wrapping Fenics. This was discussed on the discourse as probably the best short-term course of action for getting Julia a finite element toolbox. It would be at the level of a GSoC student since it can mostly be done with PyCall and, since it's just a wrapper, intimate knowledge of finite element methods won't be needed (and students will probably learn a lot in the process!). 

I also added to the Bayesian section a mention of Klara.jl.