Fix dead links (#530)
* Switch to remark-lint

This adds a new Github action to run the remark-lint
plugin remark-lint-no-dead-urls to check for dead
links. I think this is a much better tool as compared
to the previous one as this one skips the relative-urls
and also has an option to skip localhost.

* Update links

This fixes most of the dead/changed links if not all.
Some of the links have been removed since they were not
available anymore. Most of the links have been updated
to their corresponding new links.

Resolves https://github.com/JuliaLang/www.julialang.org/issues/524