Update index.md
Added "From zero to Julia!", a series of online written lessons about Julia. My English is not the best, what I wanted to say is that in this series of lessons you can learn the basics about Julia in a way suitable to a beginner. You will also encounter more advanced topics as you progress. Any suggestion on how to edit/improve the description about the course to make it more appealing is welcome! Thank you