Update Windows PATH instructions
I installed Julia 1.4.2 recently and the directory that contains `julia.exe` looks like this: `C:\Users\username\AppData\Local\Programs\Julia\Julia-1.4.2\bin`, not  `C:\Users\username\AppData\Local\Programs\Julia-1.4.2\bin` (which is what the website currently says). Furthermore, the directions on the current page says "paste in the directory noted from the installation stage", but the directory noted from the installation page **does not include the \bin**, which is confusing, because the example path provided immediately afterwards *does include* a `\bin` at the end. I've edited the language to be a bit more clear.