Alternative HSR policy (#1280)
* Alternative HSR policy

This is an alternative to https://github.com/JuliaLang/www.julialang.org/pull/1229. After some extensive discussion on
the committers channel, the following objections were raised to the original draft proposal:

1. Such research is still somewhat rare, and despite the recent high profile incidents, it seems unwise to draft full policies without a concrete need in this community
2. Requiring notification to stewards was not universally regarded as a good idea. Everybody agreed that Stewards cannot give consent to perform HSR on behalf of the participants. While this is not what was contemplated in the original PR, the concern was that any sort of advance review would subsequently be perceived as implied approval and we do not want to put Stewards into this position.

This instead amends the CoC, with the goal of explicitly putting researchers on notice that we as a community consider such research to be research on human subjects and therefore needs to be treated as such, and if you're behaving unethically you may be banned. There was a remaining question about whether such a statement was actually required, considering that we would definitely ban people for misbehavior regardless of whether or not they were a researcher. However, I would like to think that most researchers at least intend to act ethically, so I'm hoping that an explicit notice of this sort would give those who do intend ethically a chance to examine their research and consider whether they may need to follow applicable HSR practices before proceeding in their research.

* Update community/standards.md

Co-authored-by: Dilum Aluthge <dilum@aluthge.com>

Co-authored-by: Dilum Aluthge <dilum@aluthge.com>