WIP: 10 year retrospective
Monday is the 10 year anniversary of the original Julia blogpost.
We've let it sneak up on us a bit, with no particular plan to
celebrate. I do think we should do something to mark the occasion.
This PR is my proposal to that extent.

The idea is to invite all members of the Julia community to write a quick, personal, 1 paragraph retrospective on 10 years of Julia (not limited to contributors who've been around for 10 years, newer contributors can write about why they're using Julia now and what made them start using it).

The timeline is a bit compressed since we only have three days, so my suggestion would be:

- Friday: Finish the framing, decide the layout, add a few more people's example retrospectives
- Saturday & Sunday: Let people in the community add their own by making a PR on this repo targeting this branch, which we'll merge as they come in
- Monday: Publish, accepting additional paragraphs as they come in

I've taken the liberty of drafting a foreword and adding my personal paragraph, just to set the tone and give an example of what I'm imagining. Both these texts are placeholders written at 3am, so they'll need to be improved, but hopefully they give the right idea.

Assuming we like this direction, I'll need help spreading the word and manging the submission process, but I think handling it through GitHub should be easy.