RFC: AMD ROCm (and oneAPI) higher tier?
I'm not sure who controls this @jpsamaroo, but none of the tiers (or the descriptions) apply to GPUs in a strict sense(?), i.e. not about building/compiling Julia.

So it's just about "vote of confidence", and I understand ROCm support is much improved.

Don't merged without consulting, maybe ROCm should go (one) lower and oneAPI's tier just a guess.