Move "Sponsors" into a separate top-level page
There are a large number of companies providing some sort of support to Julia.
The individual employees at these companies that made these decisions often
want to have something to link to demonstrate to their mgmt chain that their
contributions are being acknowledged, so move the "Sponsors" section out from
research and expand it out to cover the current list.

This is currently not linked from anywhere, but let's agree on the content
and then we can figure out where to link it.