Amend the language around manual downloads and make it more prominent
PR 2382 made the language around manually downloading the official Julia
binaries fairly aggressive, but perhaps unnecessarily so. The concern
that motivated that PR was that newer users were manually downloading
these binaries rather than managing their installations via JuliaUp.
There is nothing inherently wrong with that method, however, though it
is much less convenient for the user.

To make the this warning friendlier to users who know what they're doing
and have specific requirements that motivate this workflow (or otherwise
just prefer it), we can soften the language a bit and make it a bit more
specific, but also make the warning itself more prominent on the page by
putting it above the "Manual Downloads" heading.